# backend/api/openapi.yaml
openapi: "3.0.3"
info:
  title: Courier API
  version: "0.1.1"
  description: >
    REST interface for the Courier-Management project — now aligned with
    the Firestore document structure that includes pickup and destination
    addresses.
servers:
  - url: http://localhost:3000
    description: Local
  - url: https://courier-api.example.com
    description: Production
paths:
...
  /couriers:
    get:
      summary: List all couriers
      operationId: listCouriers
      responses:
        '200':
          description: A list of couriers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CourierUser'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /businesses:
    get:
      summary: List all businesses
      operationId: listBusinesses
      responses:
        '200':
          description: A list of businesses
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/BusinessUser'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
components:
...
  schemas:
    GeoPoint:
      type: object
      properties:
        lat: { type: number, format: double }
        lng: { type: number, format: double }
      required: [lat, lng]
...
    Error:
      type: object
      properties:
        message: { type: string }
    BusinessUser:
      type: object
      properties:
        id:
          type: string
        email:
          type: string
        businessName:
          type: string
        role:
          type: string
          enum: [business]
        businessAddress:
          type: string
        location:
          $ref: '#/components/schemas/GeoPoint'
        placeId:
          type: string
      required: [id, email, businessName, role, businessAddress, location]
    CourierUser:
      type: object
      properties:
        id:
          type: string
        email:
          type: string
        courierName:
          type: string
        role:
          type: string
          enum: [courier]
      required: [id, email, courierName, role]
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []